backend/requirements.txt
Flask
Flask-Cors
Flask-Mail
mysql-connector-python
PyJWT
pytz
Werkzeug
python-dotenv